--- github-api-&-web-app/README.md
@@ -1,6 +1,0 @@
-# Web App using Github API
-
-
-A simple web app via GitHub API.
-
-In this Web app User will be asked to enter Github User Name and after that All the Repos under that user name will be displayed.
