--- github-api-&-web-app/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <title>GitHub API</title>
-    <link rel="icon" type="image/png" href="assets/favicon.png">
-
-    <!-- Import the Bootstrap CSS CDN -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-
-</head>
-
-<body>
-
-    <h3 class="text-center mt-5">GitHub API</h3>
-    <form id="gitHubForm" class="form-inline mx-auto" style="width: 280px">
-        <input id="usernameInput" class="form-control mb-5" type="text" name="username" placeholder="GitHub Username">
-        <input type="submit" class="btn btn-primary ml-2 mb-5" value="Submit">
-    </form>
-    <ul id="userRepos" class="list-group mx-auto mb-5" style="width: 500px">
-
-    </ul>
-
-    <script src="app.js"></script>
-</body>
-
-</html>
