--- github-issues-export/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "github-issue-export",
-  "version": "1.2.0",
-  "description": "export github repo issues to csv",
-  "main": "index.js",
-  "author": "Mat Brady",
-  "license": "MIT",
-  "private": true,
-  "engines": {
-    "node": "8.4.0",
-    "npm": "5.0.0"
-  },
-  "dependencies": {
-    "graphql": "^15.5.1",
-    "graphql-request": "^1.8.2",
-    "json2csv": "^3.11.5",
-    "minimist": "^1.2.5"
-  }
-}
